requirements.txt
ujson
black
clicks
pyzmq
tensorflow-gpu==1.10.0
pyarrow
redis
